Added exceptions

+++ db/database.py
@@ -56,16 +56,22 @@ class MongoDbConnection:
     def update_blog(self, query:str , data: BlogSchema):
         try:
             db = self.db.get_collection("blogs")
-            db.update_one({"slug": query}, {"$set": data})
-        
+            if db.count_documents({"slug": query}, limit = 1) != 0:
+                db.update_one({"slug": query}, {"$set": data})
+            else:
+                raise Exception({"status": False, "message": "The blog is not present in the database."})
+            
         except Exception as e:
             raise {"status": False, "message": str(e)}
 
-    def delete_blog(self, query: str):
+    def delete_blog(self, query: str):    
         try:
             db = self.db.get_collection("blogs")
-            db.delete_one({"slug": query})    
-
+            if db.count_documents({"slug": query}, limit = 1) != 0:
+                db.delete_one({"slug": query}) 
+            else:
+                raise Exception({"status": False, "message": "The blog is not present in the database."})
+            
         except Exception as e:
             raise {"status": False, "message": str(e)}
